Move enum pack_file_id to common/enums.h
