+++ headers/types/common/enums.h
@@ -3729,6 +3729,16 @@ enum portrait_canvas_state {
     PORTRAIT_CANVAS_HIDE = 4,
 };
 
+// Correspond to a .bin pack archive
+enum pack_file_id {
+    PACK_ARCHIVE_MONSTER = 0,
+    PACK_ARCHIVE_M_ATTACK = 1,
+    PACK_ARCHIVE_M_GROUND = 2,
+    PACK_ARCHIVE_EFFECT = 3,
+    PACK_ARCHIVE_DUNGEON = 4,
+    PACK_ARCHIVE_M_LEVEL = 5,
+};
+
 enum wan_source_type {
     WAN_SOURCE_NULL = 0,
     WAN_SOURCE_FILE = 1, // Directly loaded from a .wan
